[IE TESTS] Move ROI SLT class to the shared lib (#4198)
* [IE TESTS] Move ROI SLT class to the shared lib

* [IE Tests] Fix Ci

* Fix ci
